// front-end/src/modules/theme/theme.interfaces.tsx
export enum EThemes{
    DARK = "dark",
    LIGHT = "light"
}

export enum EThemeVariables{
    BACKGROUNDPC = "background-primary-color",
    BUTTONPC = "button-primary-color",
    FONTPC = "font-primary-color",
    FONTSC = "font-secondary-color",
    ICONSPC = "icons-primary-color",
    ICONSSC = "icons-secondary-color",
    SWITCHPC = "switch-primary-color",
    SWITCHSC = "switch-secondary-color",
}